programador-lhama/requirements.txt
blinker==1.6.2
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
filelock==3.12.4
Flask==2.3.3
greenlet==2.0.2
identify==2.5.29
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.10.0
pre-commit==3.4.0
PyYAML==6.0.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
virtualenv==20.24.5
Werkzeug==2.3.7
